convert with expect if allowed

+++ mach/video_test.go
@@ -8,5 +8,5 @@ import (
 
 func TestFitInWindow(t *testing.T) {
 	frame := FitInWindow(1024, 768, 224, 288)
-	With(t).Expect(frame.Scale).ToBe(int32(2))
+	With(t).Expect(frame.Scale).ToBe(2)
 }
